requirements/production.txt
-r base.txt
django-environ==0.4.5
gunicorn==19.9.0

mysqlclient
django-debug-toolbar
